I hate bugs. I just hate them and the worst part is I know I made the bug. Still I made some progress.

+++ zaap/src/scene/terrain/Terrain.cpp
@@ -14,7 +14,7 @@ namespace zaap { namespace scene {
 	{
 	}
 
-	TERRAIN_DESC::TERRAIN_DESC(float heightMin, float heightMax, float defaultHeight, float meshSize, float verticesPerLine)
+	TERRAIN_DESC::TERRAIN_DESC(float heightMin, float heightMax, float defaultHeight, float meshSize, uint verticesPerLine)
 		: HeightMin(heightMin),
 		HeightMax(heightMax),
 		DefaultHeight(defaultHeight),
